requirements.txt
aiohttp==3.7.4.post0
astroid==2.8.4
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
autopep8==1.5.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
boto3==1.21.40
botocore==1.24.40
certifi==2021.10.8
cffi==1.14.5
chainmap==1.0.3
chardet==4.0.0
click==8.0.3
colorama==0.4.4
combomethod==1.0.12
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
decorator==4.4.2
discord==1.7.3
discord.py==1.7.3
distlib==0.3.3
filelock==3.3.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.9.3
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
mysql==0.0.3
mysqlclient==2.0.3
nest-asyncio==1.5.1
nulltype==2.3.1
options==1.4.10
parsel==1.6.0
parso==0.8.1
pathspec==0.9.0
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.4.0
priority==1.3.0
prompt-toolkit==3.0.16
Protego==0.1.16
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.4
pylint==2.11.1
pymongo==3.12.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.20.0
python-env==1.0.0
pywin32==300
pyzmq==22.0.3
queuelib==1.6.1
regex==2021.10.23
s3transfer==0.5.2
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
six==1.12.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.8.1
virtualenv-clone==0.5.7
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.6.3
zope.interface==5.4.0
